cleanup

+++ nodemanager/monitoring.go
@@ -52,7 +52,7 @@ func (s *NodeosSuperviser) Monitor() {
 		s.chainID = chainInfo.ChainID
 		s.serverVersion = chainInfo.ServerVersion
 		s.serverVersionString = chainInfo.ServerVersionString
-		s.lastBlockSeen = uint32(chainInfo.HeadBlockNum)
+		s.lastBlockSeen = chainInfo.HeadBlockNum
 
 		lastHeadBlockTime = chainInfo.HeadBlockTime.Time
 
@@ -67,18 +67,6 @@ func (s *NodeosSuperviser) Monitor() {
 				s.Logger.Error("failed to update head block", zap.Error(err))
 			}
 		}
-
-		// monitor if BP is producer (should be 1 and only 1)
-		//if s.IsActiveProducer() {
-		//	_, err := s.api.IsProducerPaused(context.Background())
-		//	// isProducerPaused, err := s.api.IsProducerPaused(context.Background())
-		//	if err != nil {
-		//		s.Logger.Debug("unable to check if producer is paused", zap.Error(err))
-		//	} else {
-		//		// todo reimplement
-		//		// metrics.SetNodeosIsBlockProducer(isProducerPaused)
-		//	}
-		//}
 
 		if lastDbSizeTime.IsZero() || time.Now().Sub(lastDbSizeTime).Seconds() > 30.0 {
 			s.Logger.Debug("first monitoring call or more than 30s has elapsed since last call, querying db size from nodeos")
